Rename queue compute policy to delayed and add delayed provider alias
- Rename queueComputePolicy files/classes to delayedComputePolicy / DelayedComputePolicy
- Update imports, tests, and tsconfig includes to new names
- Add upgrade.service.delayed alias to delayed recompute factory
- Keep upgradeQueue for backward compatibility (deprecated in typings)
