+++ tsconfig.json
@@ -183,8 +183,8 @@
     "providers/upgrade/computePolicy.d.ts",
     "providers/upgrade/onDemandComputePolicy.d.ts",
     "providers/upgrade/onDemandComputePolicy.js",
-    "providers/upgrade/queueComputePolicy.d.ts",
-    "providers/upgrade/queueComputePolicy.js",
+    "providers/upgrade/delayedComputePolicy.d.ts",
+    "providers/upgrade/delayedComputePolicy.js",
     "providers/upgrade/recomputeHandler.d.ts",
     "providers/upgrade/recomputeHandler.js",
     "providers/upgrade/process.d.ts",
